refactor: update imports for Python 3 compatibility and improve integer calculations

+++ gui.py
@@ -1,6 +1,8 @@
+import sys
+
 try:
-    from Tkinter import *
-    import tkFont
+    from tkinter import *
+    from tkinter import font as tkFont
 except ImportError as err:
     print("error: %s. Tkinter library is required for using the GUI." % err.message)
     sys.exit(1)
@@ -10,7 +12,7 @@ from AutomataTheory import *
 dotFound = isInstalled("dot")
 if dotFound:
     try:
-        import Image, ImageTk
+        from PIL import Image, ImageTk
     except ImportError as err:
         print("Notice: %s. The PIL library is required for displaying the graphs." % err.message)
         dotFound = False
@@ -37,8 +39,8 @@ class AutomataGUI:
         self.FrameSizeX  = int(ScreenSizeX * ScreenRatioX)
         self.FrameSizeY  = int(ScreenSizeY * ScreenRatioY)
         print(self.FrameSizeY, self.FrameSizeX)
-        FramePosX   = (ScreenSizeX - self.FrameSizeX)/2
-        FramePosY   = (ScreenSizeY - self.FrameSizeY)/2
+        FramePosX = int((ScreenSizeX - self.FrameSizeX)/2)
+        FramePosY = int((ScreenSizeY - self.FrameSizeY)/2)
         padX = 10
         padY = 10
         self.root.geometry("%sx%s+%s+%s" % (self.FrameSizeX,self.FrameSizeY,FramePosX,FramePosY))
